moved checkMaxLength and findMax to "devtools" dir
