+++ devtools/findMax.ts
@@ -1,3 +1,4 @@
+//finds maximum value in a number[]
 export function findMax(arr: Array<number>): number {
     let max = arr[0];
     for (let i = 1; i < arr.length; i++) {
