+++ tests/checkMaxLength.test.ts
@@ -1,7 +1,7 @@
-import {CheckMoviePropertiesMaxLength} from "../utils/checkMoviePropertiesMaxLength";
+import {CheckMoviePropertiesMaxLength} from "../devtools/checkMoviePropertiesMaxLength";
 
 test("check max length of all properties test", async () => {
     //CheckMoviePropertiesMaxLength should not throw
     expect(async () => await CheckMoviePropertiesMaxLength()).not.toThrow();
-    console.log(await CheckMoviePropertiesMaxLength());
+    // console.log(await CheckMoviePropertiesMaxLength());
 })
